+++ extensions/music/src/main/java/app/morphe/extension/music/patches/spoof/SpoofVideoStreamsPatch.java
@@ -1,7 +1,7 @@
 package app.morphe.extension.music.patches.spoof;
 
 import static app.morphe.extension.music.settings.Settings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE;
-import static app.morphe.extension.shared.spoof.ClientType.ANDROID_REEL;
+import static app.morphe.extension.shared.spoof.ClientType.ANDROID_REEL_NO_AUTH;
 import static app.morphe.extension.shared.spoof.ClientType.ANDROID_VR_1_64;
 import static app.morphe.extension.shared.spoof.ClientType.TV;
 import static app.morphe.extension.shared.spoof.ClientType.VISIONOS;
@@ -21,7 +21,7 @@ public class SpoofVideoStreamsPatch {
                 TV,
                 ANDROID_VR_1_64,
                 VISIONOS,
-                ANDROID_REEL
+                ANDROID_REEL_NO_AUTH
         );
 
         app.morphe.extension.shared.spoof.SpoofVideoStreamsPatch.setClientsToUse(