+++ extensions/shared-youtube/library/src/main/java/app/morphe/extension/shared/settings/preference/ForceOriginalAudioSwitchPreference.java
@@ -16,7 +16,7 @@ public class ForceOriginalAudioSwitchPreference extends SwitchPreference {
     // Spoof stream patch is not included, or is spoofing to a client with multiple audio streams.
     private static final boolean available = !SpoofVideoStreamsPatch.isPatchIncluded()
             || !SharedYouTubeSettings.SPOOF_VIDEO_STREAMS.get()
-            || SpoofVideoStreamsPatch.getPreferredClient() == ClientType.ANDROID_REEL
+            || SpoofVideoStreamsPatch.getPreferredClient() == ClientType.ANDROID_REEL_AUTH
             || SpoofVideoStreamsPatch.getPreferredClient() == ClientType.TV;
 
     {