+++ extensions/youtube/src/main/java/app/morphe/extension/youtube/patches/spoof/SpoofVideoStreamsPatch.java
@@ -45,8 +45,8 @@ public class SpoofVideoStreamsPatch {
         // Reels can take up to 1 minute for videos start playback.
         // Only use it if the user has selected it.
         List<ClientType> availableClients = List.of(
-                TV,
                 ANDROID_CREATOR,
+                TV,
                 ANDROID_VR_1_64,
                 VISIONOS
         );