+++ patches/src/main/kotlin/app/morphe/patches/reddit/misc/fix/signature/SpoofSignaturePatch.kt
@@ -7,6 +7,7 @@
 package app.morphe.patches.reddit.misc.fix.signature
 
 import app.morphe.patcher.patch.bytecodePatch
+import app.morphe.patches.all.misc.fix.changepackageinstaller.changePackageInstallerPatch
 import app.morphe.patches.reddit.misc.extension.sharedExtensionPatch
 import app.morphe.patches.reddit.shared.Constants.COMPATIBILITY_REDDIT
 
@@ -20,7 +21,7 @@ val spoofSignaturePatch = bytecodePatch(
 ) {
     compatibleWith(COMPATIBILITY_REDDIT)
 
-    dependsOn(sharedExtensionPatch)
+    dependsOn(sharedExtensionPatch, changePackageInstallerPatch())
 
     execute {
         ApplicationFingerprint.classDef.setSuperClass(EXTENSION_CLASS)