+++ patches/src/main/kotlin/app/morphe/patches/shared/misc/fix/bitmap/FixRecycledBitmapPatch.kt
@@ -11,8 +11,8 @@ import app.morphe.patcher.Fingerprint
 import app.morphe.patcher.extensions.InstructionExtensions.replaceInstruction
 import app.morphe.patcher.methodCall
 import app.morphe.patcher.patch.bytecodePatch
-import app.morphe.util.findInstructionIndicesReversedOrThrow
 import app.morphe.util.fiveRegisters
+import app.morphe.util.matchAllMethodIndicesForEach
 
 private const val EXTENSION_CLASS =
     "Lapp/morphe/extension/shared/patches/FixRecycledBitmapPatch;"
@@ -22,30 +22,26 @@ val fixRecycledBitmapPatch = bytecodePatch(
 ) {
 
     execute {
-        val putBitmapCall = methodCall(
-            definingClass = $$"Landroid/media/MediaMetadata$Builder;",
-            name = "putBitmap",
-            parameters = listOf("Ljava/lang/String;", "Landroid/graphics/Bitmap;")
-        )
-
         Fingerprint(
-            filters = listOf(putBitmapCall),
+            filters = listOf(
+                methodCall(
+                    definingClass = $$"Landroid/media/MediaMetadata$Builder;",
+                    name = "putBitmap",
+                    parameters = listOf("Ljava/lang/String;", "Landroid/graphics/Bitmap;")
+                )
+            ),
             custom = { _, classDef ->
                 !classDef.type.startsWith("Lapp/morphe/extension")
             }
-        ).matchAllOrNull()?.forEach { match ->
-            match.method.apply {
-                findInstructionIndicesReversedOrThrow(putBitmapCall).forEach { index ->
-                    val registers = fiveRegisters(index)
+        ).matchAllMethodIndicesForEach(requireMatches = false) { index ->
+            val registers = fiveRegisters(index)
 
-                    replaceInstruction(
-                        index,
-                        $"invoke-static { $registers }, $EXTENSION_CLASS->putBitmap(" +
-                                "Landroid/media/MediaMetadata\$Builder;Ljava/lang/String;Landroid/graphics/Bitmap;)" +
-                                "Landroid/media/MediaMetadata\$Builder;"
-                    )
-                }
-            }
+            replaceInstruction(
+                index,
+                $"invoke-static { $registers }, $EXTENSION_CLASS->putBitmap(" +
+                        "Landroid/media/MediaMetadata\$Builder;Ljava/lang/String;Landroid/graphics/Bitmap;)" +
+                        "Landroid/media/MediaMetadata\$Builder;"
+            )
         }
     }
 }