+++ patches/src/main/kotlin/app/morphe/patches/youtube/layout/livering/OpenChannelOfLiveAvatarPatch.kt
@@ -3,11 +3,7 @@ package app.morphe.patches.youtube.layout.livering
 import app.morphe.patcher.extensions.InstructionExtensions.addInstruction
 import app.morphe.patcher.extensions.InstructionExtensions.addInstructionsWithLabels
 import app.morphe.patcher.patch.bytecodePatch
-import app.morphe.patcher.patch.resourcePatch
-import app.morphe.patches.all.misc.resources.AppLocale
 import app.morphe.patches.all.misc.resources.addResourcesPatch
-import app.morphe.patches.all.misc.resources.createResourceDestinationDirectoryIfNeeded
-import app.morphe.patches.all.misc.resources.localesYouTube
 import app.morphe.patches.shared.misc.settings.preference.SwitchPreference
 import app.morphe.patches.youtube.layout.shortsplayer.ShortsPlaybackIntentFingerprint
 import app.morphe.patches.youtube.layout.shortsplayer.ShortsPlaybackIntentFingerprintLegacy
@@ -19,38 +15,8 @@ import app.morphe.patches.youtube.shared.Constants.COMPATIBILITY_YOUTUBE
 import app.morphe.patches.youtube.shared.YouTubeActivityOnCreateFingerprint
 import app.morphe.patches.youtube.video.information.PlaybackStartDescriptorToStringFingerprint
 import app.morphe.util.addInstructionsAtControlFlowLabel
-import app.morphe.util.copyXmlNode
 import app.morphe.util.getFreeRegisterProvider
-import app.morphe.util.inputStreamFromBundledResource
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
-import java.util.logging.Logger
-
-private val openChannelOfLiveAvatarResourcePatch = resourcePatch(
-    description = "openChannelOfLiveAvatarResourcePatch"
-) {
-    execute {
-        val logger = Logger.getLogger(AppLocale.Companion::class.java.name)
-
-        localesYouTube.filter { it.isBuiltInLanguage }.forEach { locale ->
-            val directory = locale.getDestLocaleFolderName()
-            val targetResource = "$directory/strings.xml"
-            val destinationPath = "res/$targetResource"
-            val destFile = this[destinationPath]
-
-            createResourceDestinationDirectoryIfNeeded(locale, logger, destinationPath, destFile)
-
-            inputStreamFromBundledResource(
-                "livering/host",
-                targetResource
-            )!!.let { inputStream ->
-                "resources".copyXmlNode(
-                    document(inputStream),
-                    document(destinationPath)
-                ).close()
-            }
-        }
-    }
-}
 
 private const val EXTENSION_CLASS =
     "Lapp/morphe/extension/youtube/patches/OpenChannelOfLiveAvatarPatch;"
@@ -63,7 +29,6 @@ val openChannelOfLiveAvatarPatch = bytecodePatch(
     compatibleWith(COMPATIBILITY_YOUTUBE)
 
     dependsOn(
-        openChannelOfLiveAvatarResourcePatch,
         addResourcesPatch,
         settingsPatch,
         versionCheckPatch,