+++ patches/src/main/kotlin/app/morphe/patches/youtube/misc/playservice/VersionCheckPatch.kt
@@ -47,6 +47,8 @@ var is_21_02_or_greater : Boolean by Delegates.notNull()
     private set
 var is_21_03_or_greater : Boolean by Delegates.notNull()
     private set
+var is_21_04_or_greater : Boolean by Delegates.notNull()
+    private set
 var is_21_05_or_greater : Boolean by Delegates.notNull()
     private set
 var is_21_06_or_greater : Boolean by Delegates.notNull()
@@ -102,7 +104,8 @@ val versionCheckPatch = bytecodePatch {
         is_20_49_or_greater = isEqualsOrGreaterThan("20.49.00")
         is_21_02_or_greater = isEqualsOrGreaterThan("21.02.000")
         is_21_03_or_greater = isEqualsOrGreaterThan("21.03.000")
-        is_21_05_or_greater = isEqualsOrGreaterThan("20.05.000")
+        is_21_04_or_greater = isEqualsOrGreaterThan("21.04.000")
+        is_21_05_or_greater = isEqualsOrGreaterThan("21.05.000")
         is_21_06_or_greater = isEqualsOrGreaterThan("21.06.000")
         is_21_07_or_greater = isEqualsOrGreaterThan("21.07.000")
         is_21_08_or_greater = isEqualsOrGreaterThan("21.08.000")