+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+## [1.29.0](https://github.com/MorpheApp/morphe-patches/compare/v1.28.0...v1.29.0) (2026-05-21)
+
+### 🐛 Bug Fixes
+
+* **Reddit:** Spoof installation source ([e62b180](https://github.com/MorpheApp/morphe-patches/commit/e62b1803a34ffc3045b93ee5669e0d3441d0a8c1))
+* **YouTube - Open channel of live avatar:** deprecate match with translations ([#1511](https://github.com/MorpheApp/morphe-patches/issues/1511)) ([4264071](https://github.com/MorpheApp/morphe-patches/commit/42640710c8be1b173984a0d052891006960bcbcc))
+* **YouTube - Open channel of live avatar:** Fixed null exception after tap on some shorts player buttons ([#1497](https://github.com/MorpheApp/morphe-patches/issues/1497)) ([6ec2fd4](https://github.com/MorpheApp/morphe-patches/commit/6ec2fd4f8d9e77b8adffb6b91da5ebb61413f58c))
+* **YouTube - PlayerOverlayButton:** Improved containers handling ([1580f02](https://github.com/MorpheApp/morphe-patches/commit/1580f02a0dfe0a75f9f636229bc6d77cab4b3699))
+* **YouTube - Remove background playback restrictions:** Disable client flag that interferes with player overlay buttons ([f24cae5](https://github.com/MorpheApp/morphe-patches/commit/f24cae55007dd00bc061dadcaeb2c2f74baf3816))
+
+### ✨ New Features
+
+* **Reddit:** Add experimental support for `2026.20.0` ([0e6b305](https://github.com/MorpheApp/morphe-patches/commit/0e6b3054b011ac89d99dc7088dc983a1f95e99a4))
+
+## [1.29.0-dev.3](https://github.com/MorpheApp/morphe-patches/compare/v1.29.0-dev.2...v1.29.0-dev.3) (2026-05-20)
+
+### 🐛 Bug Fixes
+
+* **YouTube - Open channel of live avatar:** deprecate match with translations ([#1511](https://github.com/MorpheApp/morphe-patches/issues/1511)) ([4264071](https://github.com/MorpheApp/morphe-patches/commit/42640710c8be1b173984a0d052891006960bcbcc))
+
+## [1.29.0-dev.2](https://github.com/MorpheApp/morphe-patches/compare/v1.29.0-dev.1...v1.29.0-dev.2) (2026-05-19)
+
+### 🐛 Bug Fixes
+
+* **YouTube - PlayerOverlayButton:** Improved containers handling ([1580f02](https://github.com/MorpheApp/morphe-patches/commit/1580f02a0dfe0a75f9f636229bc6d77cab4b3699))
+* **YouTube - Remove background playback restrictions:** Disable client flag that interferes with player overlay buttons ([f24cae5](https://github.com/MorpheApp/morphe-patches/commit/f24cae55007dd00bc061dadcaeb2c2f74baf3816))
+
+## [1.29.0-dev.1](https://github.com/MorpheApp/morphe-patches/compare/v1.28.0...v1.29.0-dev.1) (2026-05-18)
+
+### 🐛 Bug Fixes
+
+* **Reddit:** Spoof installation source ([e62b180](https://github.com/MorpheApp/morphe-patches/commit/e62b1803a34ffc3045b93ee5669e0d3441d0a8c1))
+* **YouTube - Open channel of live avatar:** Fixed null exception after tap on some shorts player buttons ([#1497](https://github.com/MorpheApp/morphe-patches/issues/1497)) ([6ec2fd4](https://github.com/MorpheApp/morphe-patches/commit/6ec2fd4f8d9e77b8adffb6b91da5ebb61413f58c))
+
+### ✨ New Features
+
+* **Reddit:** Add experimental support for `2026.20.0` ([0e6b305](https://github.com/MorpheApp/morphe-patches/commit/0e6b3054b011ac89d99dc7088dc983a1f95e99a4))
+
 ## [1.28.0](https://github.com/MorpheApp/morphe-patches/compare/v1.27.0...v1.28.0) (2026-05-17)
 
 ### 🐛 Bug Fixes